\name{coursera-fars-package}
\alias{coursera-fars-package}
\alias{coursera-fars}
\docType{package}
\title{
\packageTitle{coursera-fars}
}
\description{
\packageDescription{coursera-fars}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{coursera-fars}
\packageIndices{coursera-fars}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{coursera-fars}

Maintainer: \packageMaintainer{coursera-fars}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
